requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
cryptography>=42.0.0
python-dotenv>=1.0.0
requests>=2.31.0
bcrypt>=4.1.0
claude-agent-sdk>=0.1.69
pydantic>=2.7.0
python-multipart>=0.0.9
